Renamed SimpleTable to SectionSimpleTable
I renamed the file because its just a section for the styleguide
