+++ frontend/src/styleguide/SectionSimpleTable.vue
