+++ frontend/src/views/Styleguide.vue
@@ -48,7 +48,7 @@
       </styleguide-section>
 
       <styleguide-section title="Simple-Table">
-        <SimpleTable />
+        <section-simple-table />
       </styleguide-section>
     </div>
   </div>
@@ -63,7 +63,7 @@ import StyleguideIcons from '@/styleguide/Icons';
 import SectionButtons from '@/styleguide/SectionButtons';
 import SectionTooltip from '@/styleguide/SectionTooltip';
 import ForgeToggle from '@/components/Toggle';
-import SimpleTable from '@/styleguide/SimpleTable';
+import SectionSimpleTable from '@/styleguide/SectionSimpleTable';
 
 export default {
   components: {
@@ -75,7 +75,7 @@ export default {
     StyleguideFormElements,
     SectionTooltip,
     ForgeToggle,
-    SimpleTable,
+    SectionSimpleTable,
   },
 };
 </script>
